requirements.txt
numpy
scipy
Pillow
tqdm
